[Validation] Moving Validate classes to a new Validator namespace for better organization.
